rename popover menu file
