+++ assets/src/dashboard/components/index.js
@@ -28,7 +28,7 @@ export {
 export { default as Dropdown } from './dropdown';
 export { default as NavigationBar } from './navigationBar';
 export { Pill, FloatingTab } from './pill';
-export { default as PopoverMenu } from './popover-menu';
+export { default as PopoverMenu } from './popoverMenu';
 export { default as PopoverPanel } from './popoverPanel';
 export { default as PreviewPage } from './previewPage';
 export { default as TypeaheadInput } from './typeaheadInput';
